Separating controllers from routes
