--- entity.json
@@ -1,35 +1,0 @@
-{
-    "user" :{
-        "id" : 1,
-        "identifications" : {
-            "email" : "fezzopro@gmail.com",
-            "first_name" : "Felix",
-            "last_name" : "KAMANA",
-            "passwor" : "password",
-            "dob" : "2019-07-20",
-            "is_admin" : "true"
-        }
-    },
-    "trip": {
-        "id" : "1",
-        "identification": {
-            "seating_capacity" : "20",
-            "bus_license_number" : "RAD787D",
-            "origin" : "Kigali", 
-            "destination" : "Kibungo",
-            "trip_date" : "2019-07-20",
-            "fare" : "2500.25",
-            "status" : "Active"
-            }  
-    },
-    "Booking":{
-        "id":1,
-        "identification":{
-            "trip_id": "1",
-            "user_id": 1,
-            "created_on": "2019-07-20",
-            "seat_number": "7"
-        }
-    }
-
-}
